Fix issues with buffer handling

+++ Client.h
@@ -39,7 +39,7 @@ private:
     void handleCommandFailNext(FailNext *command);
 
     int sendWriteMessage(size_t serverIdx, int nextVersion, const std::string &filename, const std::string &content);
-    std::optional<std::tuple<int, std::string>> sendReadMessage(size_t serverIdx, const std::string &filename);
+    std::optional<std::tuple<int, std::string>> sendReadMessage(int serverIdx, const std::string &filename);
     int sendGetVersion(size_t serverIdx, const std::string &filename);
     void sendFailNext(size_t serverIdx);
